BOX-106: refactor(card-draft): add draft-form
