+++ src/features/card/draft/README.md
@@ -1,4 +1,6 @@
-# Набор draft-действий с карточкой
+# Форма черновика карточки
+
+Контролы:
 
 - `ResetChanges` - сбросить изменения по карточке (+ сброс черновика)
 - `SubmitChanges` - отправить изменения по карточке (+ сброс черновика)
