+++ src/features/card/draft/index.ts
@@ -1,2 +1,2 @@
-export * from './ui';
+export * as CardDraft from './ui';
 export * as cardDraftModel from './model';
