+++ src/features/card/draft/ui/atoms/reset-changes.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { button } from '@box/ui';
 import { useEvent } from 'effector-react/ssr';
 
-import * as model from '../model';
+import * as model from '../../model';
 
 // eslint-disable-next-line prettier/prettier
 const CANCEL_WARN = 'Are you sure you want to undo the changes? The action is not reversible!';
