+++ src/features/card/draft/ui/atoms/submit-changes.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { button } from '@box/ui';
 import { useEvent, useStore } from 'effector-react/ssr';
 
-import * as model from '../model';
+import * as model from '../../model';
 
 interface Props {
   title?: string;
