+++ src/features/card/draft/ui/index.ts
@@ -1,6 +1,2 @@
-// FIXME: maybe decompose later (to separate features)
-
-export * from './submit-changes';
-export * from './reset-changes';
-export * from './edit-title';
-export * from './edit-content';
+export * from './atoms';
+export * from './molecules';
