+++ src/pages/card/create/index.tsx
@@ -1,7 +1,7 @@
-import * as CardDraft from '@box/features/card/draft';
 import React from 'react';
 import styled from 'styled-components';
-import { ContentCenteredTemplate, button } from '@box/ui';
+import { CardDraft } from '@box/features/card/draft';
+import { ContentCenteredTemplate } from '@box/ui';
 import { Helmet } from 'react-helmet-async';
 import { useStart, withStart } from '@box/lib/page-routing';
 
@@ -18,18 +18,7 @@ export const CardCreatePage = () => {
       <Helmet title="Create card" />
       <ContentCenteredTemplate>
         <Container>
-          <Header>
-            <CardDraft.EditTitle />
-          </Header>
-          <Content>
-            <CardDraft.EditContent />
-          </Content>
-          <Footer>
-            <button.Group>
-              <CardDraft.SubmitChanges title="Create and publish" />
-              <CardDraft.ResetChanges />
-            </button.Group>
-          </Footer>
+          <CardDraft.Form okText="Create and publish" />
         </Container>
       </ContentCenteredTemplate>
     </>
@@ -40,17 +29,4 @@ withStart(model.pageLoaded, CardCreatePage);
 
 const Container = styled.div`
   margin: 30px 120px 120px 120px;
-`;
-
-// Layout
-const Header = styled.div`
-  margin-bottom: 50px;
-`;
-
-const Content = styled.div`
-  min-height: 500px;
-`;
-
-const Footer = styled.div`
-  margin-top: 100px;
 `;
