+++ src/pages/card/edit/index.tsx
@@ -1,7 +1,7 @@
-import * as CardDraft from '@box/features/card/draft';
 import React from 'react';
 import styled from 'styled-components';
-import { ContentCenteredTemplate, Text, TextType, button } from '@box/ui';
+import { CardDraft } from '@box/features/card/draft';
+import { ContentCenteredTemplate, Text, TextType } from '@box/ui';
 import { Helmet } from 'react-helmet-async';
 import { combine } from 'effector-root';
 import { useStart, withStart } from '@box/lib/page-routing';
@@ -36,22 +36,7 @@ const PageContent = variant({
     // FIXME: replace to Error widget later
     // @see https://ant.design/components/result/#components-result-demo-404
     notFound: () => <Text type={TextType.header2}>Card not found</Text>,
-    ready: () => (
-      <>
-        <Header>
-          <CardDraft.EditTitle />
-        </Header>
-        <Content>
-          <CardDraft.EditContent />
-        </Content>
-        <Footer>
-          <button.Group>
-            <CardDraft.SubmitChanges title="Save" />
-            <CardDraft.ResetChanges />
-          </button.Group>
-        </Footer>
-      </>
-    ),
+    ready: () => <CardDraft.Form okText="Save" />,
   },
 });
 
@@ -59,17 +44,4 @@ withStart(model.pageLoaded, CardEditPage);
 
 const Container = styled.div`
   margin: 30px 120px 120px 120px;
-`;
-
-// Layout
-const Header = styled.div`
-  margin-bottom: 50px;
-`;
-
-const Content = styled.div`
-  min-height: 500px;
-`;
-
-const Footer = styled.div`
-  margin-top: 100px;
 `;
